requirements.txt
pandas 
numpy
tensorflow
streamlit
matplotlib
opencv-python-headless
pathlib
scikit-learn
Pillow
uuid